SONIX-requirements.txt
google-generativeai
SpeechRecognition
pyttsx3
pyaudio
python-dotenv
